+++ src/main/java/com/tickit/batch/TickitBatchApplication.java
@@ -1,0 +1,13 @@
+package com.tickit.batch;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TickitBatchApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TickitBatchApplication.class, args);
+	}
+
+}
