+++ src/main/java/com/tickit/batch/adapter/upbit/UpbitClient.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.adapter.upbit;
+
+public class UpbitClient {
+}
