+++ src/main/java/com/tickit/batch/adapter/upbit/UpbitMarketCodeAdapter.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.adapter.upbit;
+
+public class UpbitMarketCodeAdapter {
+}
