+++ src/main/java/com/tickit/batch/adapter/upbit/dto/MarketCodeResponse.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.adapter.upbit.dto;
+
+public class MarketCodeResponse {
+}
