+++ src/main/java/com/tickit/batch/adapter/upbit/dto/TickerResponse.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.adapter.upbit.dto;
+
+public class TickerResponse {
+}
