+++ src/main/java/com/tickit/batch/config/BatchInfraConfig.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.config;
+
+public class BatchInfraConfig {
+}
