+++ src/main/java/com/tickit/batch/config/DataSourceConfig.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.config;
+
+public class DataSourceConfig {
+}
