+++ src/main/java/com/tickit/batch/domain/Ticker.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.domain;
+
+public class Ticker {
+}
