+++ src/main/java/com/tickit/batch/item/processor/TickerProcessor.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.item.processor;
+
+public class TickerProcessor {
+}
