+++ src/main/java/com/tickit/batch/item/reader/TickerReader.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.item.reader;
+
+public class TickerReader {
+}
