+++ src/main/java/com/tickit/batch/item/writer/TickerWriter.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.item.writer;
+
+public class TickerWriter {
+}
