+++ src/main/java/com/tickit/batch/job/MarketCodeJobConfig.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.job;
+
+public class MarketCodeJobConfig {
+}
