+++ src/main/java/com/tickit/batch/launch/BatchJobLauncher.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.launch;
+
+public class BatchJobLauncher {
+}
