+++ src/main/java/com/tickit/batch/repository/MarketCodeRepository.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.repository;
+
+public class MarketCodeRepository {
+}
