+++ src/main/java/com/tickit/batch/repository/TickerRepository.java
@@ -1,0 +1,4 @@
+package com.tickit.batch.repository;
+
+public class TickerRepository {
+}
