--- src/main/java/com/tickit_batch/TickitBatchApplication.java
@@ -1,13 +1,0 @@
-package com.tickit_batch;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TickitBatchApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TickitBatchApplication.class, args);
-	}
-
-}
