+++ src/test/java/com/tickit/batch/TickitBatchApplicationTests.java
@@ -1,0 +1,13 @@
+package com.tickit.batch;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TickitBatchApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
