--- src/test/java/com/tickit_batch/TickitBatchApplicationTests.java
@@ -1,13 +1,0 @@
-package com.tickit_batch;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TickitBatchApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
